review: returnLink removed from checkout and thank-you, returnLink session storage accessed in thank-you
